added file input to separate file
